Android build complete not work tho

+++ V2/dsdl-engine/SceneManager.cpp
@@ -46,7 +46,7 @@ namespace DsdlEngine{
 
 	//Clean up scenes
 	void SceneManager::destroy(){
-		for (size_t i = 0; i < m_pScenes.size(); i++){
+		for (int i = 0; i < m_pScenes.size(); i++){
 			m_pScenes[i]->destroyScene();
 			delete m_pScenes[i];
 		}
